refactor: 레거시 compose DTO/엔드포인트/서비스 완전 제거 및 모던 구조로 정비, DTO 파일명 단수화 표준화
