+++ src/modules/stellar-systems/stellar-systems.service.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import {
   ForbiddenException,
   Injectable,
@@ -5,12 +6,10 @@ import {
 } from '@nestjs/common';
 import { PrismaService } from '../../prisma/prisma.service';
 import {
-  ComposeRequestDto,
   CreateStellarSystemDto,
   UpdateStellarSystemDto,
   CloneStellarSystemDto,
-} from './dto/stellar-systems.dto';
-import { PaginationDto } from '../../common/dto/pagination.dto';
+} from './dto/stellar-system.dto';
 
 @Injectable()
 export class StellarSystemService {
@@ -21,7 +20,10 @@ export class StellarSystemService {
    * - 스텔라 시스템과 항성이 동시에 생성됩니다
    * - 항성은 삭제할 수 없으며, 시스템당 정확히 하나만 존재합니다
    */
-  async createStellarSystem(userId: string, dto: CreateStellarSystemDto) {
+  async createStellarSystem(
+    userId: string,
+    dto: CreateStellarSystemDto
+  ): Promise<any> {
     // 갤럭시 존재 및 소유자 확인
     const galaxy = await this.prisma.galaxy.findUnique({
       where: { id: dto.galaxy_id },
@@ -36,7 +38,7 @@ export class StellarSystemService {
       throw new ForbiddenException('이 갤럭시에 대한 권한이 없습니다.');
     }
 
-    return await this.prisma.$transaction(async tx => {
+    const result = await this.prisma.$transaction(async (tx: any) => {
       // 1. 스텔라 시스템 생성 (스키마의 모든 필수 필드 포함)
       const system = await tx.stellarSystem.create({
         data: {
@@ -56,9 +58,6 @@ export class StellarSystemService {
         brightness: 75, // Volume 75%
         color: 60, // Key/Scale
         size: 50, // Complexity 2
-        temperature: 50,
-        luminosity: 50,
-        radius: 50,
         ...dto.star_properties, // 사용자 제공 속성으로 덮어쓰기
       };
 
@@ -93,6 +92,8 @@ export class StellarSystemService {
         planets: [], // 초기에는 행성이 없음
       };
     });
+
+    return result;
   }
 
   /**
@@ -196,6 +197,7 @@ export class StellarSystemService {
       }
 
       // 3. 행성들 복제
+
       const clonedPlanets: any[] = [];
       for (const planet of sourceSystem.planets) {
         const clonedPlanet = await tx.planet.create({
@@ -225,11 +227,16 @@ export class StellarSystemService {
         star: clonedStar
           ? {
               id: (clonedStar as any).id,
+
               system_id: (clonedStar as any).system_id,
+
               name: (clonedStar as any).name || 'CENTRAL STAR',
               object_type: 'STAR' as const,
+
               properties: (clonedStar as any).properties,
+
               created_at: (clonedStar as any).created_at,
+
               updated_at: (clonedStar as any).updated_at,
             }
           : null,
@@ -277,7 +284,6 @@ export class StellarSystemService {
       where: { id },
       data: {
         title: dto.name, // name → title 매핑
-        // description 필드가 스키마에 없으므로 제거
       },
       include: {
         star: true,
@@ -308,325 +314,6 @@ export class StellarSystemService {
     return this.prisma.stellarSystem.delete({
       where: { id },
     });
-  }
-
-  // === 기존 compose 메서드 (레거시 호환성) ===
-  /**
-   * 하나의 트랜잭션으로:
-   *  - 새 Galaxy 생성 or 기존 Galaxy 사용
-   *  - 각 System: 기존 연결 or 새로 생성
-   *  - 각 Planet: 새로 생성 (기존 연결 기능이 필요하면 connect 로직 추가 가능)
-   *  - 생성 메타는 인증 userId 기준으로 세팅
-   *  - 최종적으로 갤럭시 전체 상태(include) 리턴
-   */
-  async compose(userId: string, dto: ComposeRequestDto) {
-    return this.prisma.$transaction(async tx => {
-      // 1. Galaxy: 기존 or 새로 생성
-      let galaxyId = dto.galaxy_id;
-
-      if (galaxyId) {
-        // 기존 갤럭시 존재 확인
-        const existingGalaxy = await tx.galaxy.findUnique({
-          where: { id: galaxyId },
-          select: { id: true, owner_id: true },
-        });
-        if (!existingGalaxy) {
-          throw new NotFoundException('해당 갤럭시를 찾을 수 없습니다.');
-        }
-        // 소유자 확인
-        if (existingGalaxy.owner_id !== userId) {
-          throw new ForbiddenException('이 갤럭시에 대한 권한이 없습니다.');
-        }
-      } else {
-        // 새 갤럭시 생성
-        const newGalaxy = await tx.galaxy.create({
-          data: { name: dto.galaxy_name!, owner_id: userId },
-          select: { id: true },
-        });
-        galaxyId = newGalaxy.id;
-      }
-
-      // 2. Systems & Planets 처리
-      for (const systemDto of dto.systems) {
-        let systemId = systemDto.system_id;
-
-        if (!systemId) {
-          // 새 시스템 생성
-          const createdSystem = await tx.stellarSystem.create({
-            data: {
-              galaxy_id: galaxyId!,
-              title: systemDto.title!, // DTO 검증으로 보장
-              owner_id: userId,
-              created_by_id: userId,
-              original_author_id: userId,
-              created_via: 'MANUAL',
-              planets: systemDto.planets?.length
-                ? {
-                    create: systemDto.planets.map(p => ({
-                      name: p.name,
-                      object_type: 'PLANET',
-                      instrument_role: p.instrument_role || null,
-                      is_active: p.is_active ?? true,
-                      properties: p.properties || {},
-                    })),
-                  }
-                : undefined,
-            },
-            select: { id: true },
-          });
-          systemId = createdSystem.id;
-        } else {
-          // 기존 시스템에 행성 추가
-          const existingSystem = await tx.stellarSystem.findUnique({
-            where: { id: systemId },
-            select: { id: true, galaxy_id: true, owner_id: true },
-          });
-          if (!existingSystem) {
-            throw new NotFoundException('해당 시스템을 찾을 수 없습니다.');
-          }
-          // 소유자 및 갤럭시 일치 확인
-          if (
-            existingSystem.owner_id !== userId ||
-            existingSystem.galaxy_id !== galaxyId
-          ) {
-            throw new ForbiddenException(
-              '이 시스템에 대한 권한이 없거나 갤럭시가 일치하지 않습니다.'
-            );
-          }
-          // 행성 추가
-          if (systemDto.planets?.length) {
-            for (const planetDto of systemDto.planets) {
-              await tx.planet.create({
-                data: {
-                  system_id: systemId,
-                  name: planetDto.name,
-                  object_type: planetDto.object_type || 'PLANET',
-                  instrument_role: planetDto.instrument_role || null,
-                  is_active: planetDto.is_active ?? true,
-                  properties: planetDto.properties || {},
-                },
-              });
-            }
-          }
-        }
-      }
-      // 3) 최종 결과: 갤럭시 전체 상태 반환
-      const result = await tx.galaxy.findUnique({
-        where: { id: galaxyId! },
-        include: {
-          owner: { select: { id: true, username: true } },
-          stellar_systems: {
-            orderBy: { created_at: 'desc' },
-            include: { planets: { orderBy: { created_at: 'desc' } } },
-          },
-        },
-      });
-      return result;
-    });
-  }
-
-  /**
-   * 갤럭시 ID로 시스템들 조회
-   * pagination 적용
-   */
-  async readAllCompose(galaxyId: string, pagination: PaginationDto) {
-    // 갤럭시 존재 확인
-    const galaxy = await this.prisma.galaxy.findUnique({
-      where: { id: galaxyId },
-      select: {
-        id: true,
-        name: true,
-        owner_id: true,
-        created_at: true,
-        updated_at: true,
-      },
-    });
-    if (!galaxy) throw new NotFoundException('해당 갤럭시를 찾을 수 없습니다.');
-
-    const systems = await this.prisma.stellarSystem.findMany({
-      where: { galaxy_id: galaxyId },
-      skip: pagination.skip,
-      take: pagination.take,
-      orderBy: { created_at: 'desc' },
-      include: {
-        planets: {
-          orderBy: { created_at: 'desc' },
-          // pattern까지 필요하면 ↓
-          // include: { pattern: true },
-        },
-      },
-    });
-
-    return { ...galaxy, stellar_systems: systems };
-  }
-
-  /**
-   * 내가 만든 시스템만 조회 가능 (created_by_id 기준)
-   * 권한 없으면 403
-   */
-  async readMyAllCompose(
-    userId: string,
-    systemId: string,
-    pagination: PaginationDto
-  ) {
-    const { skip, take, page = 1, limit = 20 } = pagination;
-
-    //1. 시스템 존재 및 소유 확인
-    const system = await this.prisma.stellarSystem.findUnique({
-      where: { id: systemId },
-      select: {
-        id: true,
-        title: true,
-        owner_id: true, // 참고용
-        created_at: true,
-        updated_at: true,
-        galaxy: { select: { id: true, name: true, owner_id: true } },
-        planets: { take, skip, orderBy: { created_at: 'desc' } },
-        // pattern까지 필요하면 ↓
-        // include: { pattern: true },
-      },
-    });
-    if (!system) throw new NotFoundException('해당 시스템을 찾을 수 없습니다.');
-    if (system.owner_id !== userId) {
-      throw new ForbiddenException('이 시스템에 대한 권한이 없습니다.');
-    }
-
-    // 2) 플래닛 카운트 → 페이지 보정
-    const total = await this.prisma.planet.count({
-      where: { system_id: systemId },
-    });
-
-    // total=0 처리: 페이지는 1로 고정, 결과는 빈 배열
-    if (total === 0) {
-      return {
-        ...system,
-        planets: [],
-        pagination: { total, page: 1, limit, totalPages: 0 },
-      };
-    }
-
-    const totalPages = Math.ceil(total / limit);
-    const currentPage = Math.min(Math.max(1, page), totalPages);
-    const planetSkip = (currentPage - 1) * limit;
-
-    // 3) 플래닛 페이지네이션 조회 (필요 시 pattern 포함)
-    const planets = await this.prisma.planet.findMany({
-      where: { system_id: systemId },
-      orderBy: { created_at: 'desc' },
-      skip: planetSkip,
-      take: limit,
-      // pattern까지 필요하면 ↓
-      // ...(inc ? { include: { pattern: true } } : {}),
-    });
-
-    // 4) compose 형태로 반환
-    return {
-      ...system,
-      planets,
-      pagination: { total, page: currentPage, limit, totalPages },
-    };
-  }
-
-  /*
-   *  단일 시스템 번들 조회
-   *  - system + system.planets(+pattern)
-   *  - galaxy 메타 + (옵션) galaxy.stellar_systems(+planets)
-   *  - 인증 필요 없음
-   */
-  async readOneCompose(systemId: string, includeGalaxySystems = false) {
-    // 1) 시스템 및 행성 조회 (패턴 포함 가능)
-    const system = await this.prisma.stellarSystem.findUnique({
-      where: { id: systemId },
-      include: {
-        planets: { orderBy: { created_at: 'desc' } },
-        // pattern까지 필요하면 ↓
-        // include: { pattern: true },
-      },
-    });
-    if (!system) throw new NotFoundException('시스템을 찾을 수 없습니다.');
-    // 2) 갤럭시 메타 조회
-    const galaxy = await this.prisma.galaxy.findUnique({
-      where: { id: system.galaxy_id },
-      select: {
-        id: true,
-        name: true,
-        owner_id: true,
-        created_at: true,
-        updated_at: true,
-        ...(includeGalaxySystems
-          ? {
-              stellar_systems: {
-                orderBy: { created_at: 'desc' },
-                include: { planets: { orderBy: { created_at: 'desc' } } },
-              },
-            }
-          : {}),
-      },
-    });
-    if (!galaxy) {
-      throw new NotFoundException('연결된 갤럭시를 찾을 수 없습니다.');
-    }
-    return { ...galaxy, stellar_system: system };
-  }
-
-  /*
-   *내가 만든 시스템 단일 조회
-   * - created_by_id 기준
-   * - 권한 없으면 403
-   * - system + planets(+pattern)
-   *   - galaxy 메타 + (옵션) galaxy.stellar_systems(+planets)
-   * - 인증 필요
-   */
-  async readMyOneCompose(
-    userId: string,
-    systemId: string,
-    includeGalaxySystems = false
-  ) {
-    // 1. 시스템 존재 및 소유 확인
-    this.checkSystemOwner({ userId, systemId });
-    // 2. 시스템 및 행성 조회 (패턴 포함 가능)
-    const system = await this.prisma.stellarSystem.findUnique({
-      where: { id: systemId },
-      include: {
-        planets: { orderBy: { created_at: 'desc' } },
-        // pattern까지 필요하면 ↓
-        // include: { pattern: true },
-      },
-    });
-    if (!system) throw new NotFoundException('시스템을 찾을 수 없습니다.');
-    // 3. 갤럭시 메타 조회
-    const galaxy = await this.prisma.galaxy.findUnique({
-      where: { id: system.galaxy_id },
-      select: {
-        id: true,
-        name: true,
-        owner_id: true,
-        created_at: true,
-        updated_at: true,
-        ...(includeGalaxySystems
-          ? {
-              stellar_systems: {
-                orderBy: { created_at: 'desc' },
-                include: { planets: { orderBy: { created_at: 'desc' } } },
-              },
-            }
-          : {}),
-      },
-    });
-    if (!galaxy) {
-      throw new NotFoundException('연결된 갤럭시를 찾을 수 없습니다.');
-    }
-    return { ...galaxy, stellar_system: system };
-  }
-
-  private async checkSystemOwner({
-    userId,
-    systemId,
-  }: {
-    userId: string;
-    systemId: string;
-  }) {
-    throw new Error('해당 기능을 지원하지 않습니다');
   }
 
   /**
